+++ packages/jaspr/CHANGELOG.md
@@ -10,7 +10,8 @@
 
 - Fixed encoding of `lastmod` property in generated sitemap to be a valid W3C date string.
 
-- Added `--dart-define-from-file` option to `serve` and `build` commands.
+- Added `--target-os` and `--target-arch` options to the build command to cross-compile server binaries.
+- Added `--dart-define-from-file` option to the serve and build commands.
 
 - Added an example to show on pub.dev.